0-print_listint.c

+++ 0x12-more_singly_linked_lists/0-print_listint.c
@@ -1,7 +1,6 @@
 #include "lists.h"
 #include <stdio.h>
 #include <stdlib.h>
-#include <string.h>
 
 /**
  * print_listint - function that prints all the elements of a listint_t list.
@@ -10,13 +9,13 @@
  */
 size_t print_listint(const listint_t *h)
 {
-	size_t i = 0;
+	int i = 0;
 
 	if (h == NULL)
 		return (0);
 	while (h != NULL)
 	{
-		printf("%u\n", h->n);
+		printf("%d\n", h->n);
 	h = h->next;
 	i++;
 	}
